# docs/swagger.yaml
basePath: /
definitions:
  handler.companyReg:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
      workField:
        type: string
    type: object
  handler.customerReg:
    properties:
      birth:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  handler.getGeneral:
    properties:
      ID:
        type: integer
      amount:
        type: number
      birth:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
      workField:
        type: string
    type: object
  handler.session:
    properties:
      cookie:
        type: string
      id:
        type: integer
    type: object
  handler.work:
    properties:
      ID:
        type: integer
      companyID:
        type: integer
      companyName:
        type: string
      date:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      workField:
        type: string
    type: object
  handler.workAction:
    properties:
      workID:
        type: integer
    type: object
  handler.workCreateReq:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      workField:
        type: string
    type: object
  handler.workCreateResp:
    properties:
      ID:
        type: integer
      companyID:
        type: integer
      companyName:
        type: string
      date:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      workField:
        type: string
    type: object
  handler.workFields:
    properties:
      workFields:
        items:
          type: string
        type: array
    type: object
host: localhost:8081
info:
  contact:
    email: karshyga.beknur@gmail.com
  description: API for hiring companies by their type of work
  title: net-fix
  version: "1.0"
paths:
  /profile:
    get:
      consumes:
      - application/json
      description: getting profile by cookie
      operationId: profile
      parameters:
      - description: session
        in: header
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: general
          schema:
            $ref: '#/definitions/handler.getGeneral'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: get profile
      tags:
      - general
  /register/company:
    post:
      consumes:
      - application/json
      description: registration
      operationId: auth_company
      parameters:
      - description: registration
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.companyReg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.session'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      summary: register new customer
      tags:
      - auth
  /register/customer:
    post:
      consumes:
      - application/json
      description: registration
      operationId: auth_customer
      parameters:
      - description: registration
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.customerReg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.session'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      summary: Register new customer
      tags:
      - auth
  /service:
    post:
      consumes:
      - application/json
      description: 'choose one of work field: Air Conditioner,Carpentry,Electricity,Gardening,Home
        Machines,Housekeeping,Interior Design,Locks,Painting,Plumbing,Water Heaters'
      operationId: create work
      parameters:
      - description: service
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.workCreateReq'
      - description: session
        in: header
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.workCreateResp'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: create new service/work
      tags:
      - service
  /service/{id}:
    get:
      consumes:
      - application/json
      operationId: get work
      produces:
      - application/json
      responses:
        "200":
          description: okey
          schema:
            $ref: '#/definitions/handler.work'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      summary: get work by id
      tags:
      - service
  /service/apply:
    post:
      consumes:
      - application/json
      operationId: apply for work
      parameters:
      - description: work
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.workAction'
      - description: session
        in: header
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OKEY
          schema:
            type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: apply for service
      tags:
      - service
  /service/finish:
    post:
      consumes:
      - application/json
      operationId: finish work
      parameters:
      - description: finished
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.workAction'
      - description: session
        in: header
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OKEY
          schema:
            type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: finish work
      tags:
      - service
  /workfields:
    get:
      consumes:
      - application/json
      operationId: work fields
      produces:
      - application/json
      responses:
        "200":
          description: okey
          schema:
            $ref: '#/definitions/handler.workFields'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server Error
      summary: get all work fields
      tags:
      - service
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: session
    type: apiKey
swagger: "2.0"